+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### 0.3.0
+  * Added `targets batch` command, which allows for testing multiple submisssions at once.
+  * Improved documentation
+  * Submissions can now be viewed [here](http://darttargets.com/results)
+
 #### 0.2.3
   * Added `IOTarget`, which allows testing of other languages. For an example in Java, download the `java-example` assignment with `targets get` or `targets init`